--- internal/server/error.go
@@ -1,12 +1,0 @@
-package server
-
-import (
-	"log"
-
-	"github.com/awnumar/memguard"
-)
-
-func errorAndSafeExit(s string, c int) {
-	log.Println(s)
-	memguard.SafeExit(c)
-}
